+++ tests/Unit/Service/JobServiceTest.php
@@ -95,11 +95,12 @@ class JobServiceTest extends TestCase {
         $languages = [
                 'any'
         ];
+        $replace = true;
         $this->l10nMock->expects($this->once())
             ->method('t')
             ->with('Empty parameters passed.')
             ->will($this->returnValue('Empty parameters passed.'));
-        $this->cut->process($languages, $files);
+        $this->cut->process($languages, $files, $replace);
     }
 
     /**
@@ -113,6 +114,7 @@ class JobServiceTest extends TestCase {
         $languages = [
                 'fra'
         ];
+        $replace = true;
         $this->appConfigServiceMock->expects($this->once())
             ->method('getAppValue')
             ->with('languages')
@@ -121,7 +123,7 @@ class JobServiceTest extends TestCase {
             ->method('t')
             ->with('Empty parameters passed.')
             ->will($this->returnValue('Empty parameters passed.'));
-        $this->cut->process($languages, $files);
+        $this->cut->process($languages, $files, $replace);
     }
 
     /**
@@ -135,6 +137,7 @@ class JobServiceTest extends TestCase {
         $languages = [
                 ''
         ];
+        $replace = true;
         $this->appConfigServiceMock->expects($this->once())
             ->method('getAppValue')
             ->with('languages')
@@ -143,7 +146,62 @@ class JobServiceTest extends TestCase {
             ->method('t')
             ->with('Empty parameters passed.')
             ->will($this->returnValue('Empty parameters passed.'));
-        $this->cut->process($languages, $files);
+        $this->cut->process($languages, $files, $replace);
+    }
+
+    /**
+     * @expectedException OCA\Ocr\Service\NotFoundException
+     * @expectedExceptionMessage Empty parameters passed.
+     */
+    public function testProcessWrongReplaceType() {
+        $files = [
+                'id' => 1
+        ];
+        $languages = [
+                'deu'
+        ];
+        $replace = '';
+        $this->appConfigServiceMock->expects($this->once())
+            ->method('getAppValue')
+            ->with('languages')
+            ->will($this->returnValue('deu;eng;spa;deu-frak'));
+        $this->l10nMock->expects($this->once())
+            ->method('t')
+            ->with('Empty parameters passed.')
+            ->will($this->returnValue('Empty parameters passed.'));
+        $this->cut->process($languages, $files, $replace);
+    }
+
+    /**
+     * @expectedException OCA\Ocr\Service\NotFoundException
+     * @expectedExceptionMessage Cannot replace shared files.
+     */
+    public function testProcessFailsWithSharedAndReplace() {
+        $files = [
+                'id' => 2
+        ];
+        $languages = [
+                'deu'
+        ];
+        $replace = true;
+        $this->appConfigServiceMock->expects($this->once())
+            ->method('getAppValue')
+            ->with('languages')
+            ->will($this->returnValue('deu;eng;spa;deu-frak'));
+        $this->fileServiceMock->expects($this->once())
+            ->method('buildFileInfo')
+            ->with($files)
+            ->will($this->returnValue([$this->fileInfoSharedPdf]));
+        // foreach from here on
+        $this->fileServiceMock->expects($this->once())
+            ->method('checkSharedWithInitiator')
+            ->with($this->fileInfoSharedPdf)
+            ->will($this->returnValue(true));
+        $this->l10nMock->expects($this->once())
+            ->method('t')
+            ->with('Cannot replace shared files.')
+            ->will($this->returnValue('Cannot replace shared files.'));
+        $result = $this->cut->process($languages, $files, $replace);
     }
 
     public function testProcessSpecificNoLanguage() {
@@ -161,6 +219,7 @@ class JobServiceTest extends TestCase {
         $languages = [
                 'any'
         ];
+        $replace = false;
         $this->appConfigServiceMock->expects($this->once())
             ->method('getAppValue')
             ->with('languages')
@@ -188,16 +247,22 @@ class JobServiceTest extends TestCase {
             ->will($this->onConsecutiveCalls(false, true, true));
         $this->fileServiceMock->expects($this->exactly(3))
             ->method('buildTarget')
-            ->withConsecutive([
-                $this->fileInfoNotSharedPdf,
-                false
-        ], [
-                $this->fileInfoSharedPdf,
-                true
-        ], [
-                $this->fileInfoSharedPng,
-                true
-        ])
+            ->withConsecutive(
+                [
+                        $this->fileInfoNotSharedPdf,
+                        false,
+                        false
+                ], 
+                [
+                        $this->fileInfoSharedPdf,
+                        true,
+                        false
+                ], 
+                [
+                        $this->fileInfoSharedPng,
+                        true,
+                        false
+                ])
             ->will(
                 $this->onConsecutiveCalls('/test/path/to/file_OCR.pdf', '/test/path/to/file_OCR.pdf', 
                         '/test/path/to/file_OCR.txt'));
@@ -248,13 +313,13 @@ class JobServiceTest extends TestCase {
         // step out from tempfile creation
         $job1 = new OcrJob(OcrConstants::STATUS_PENDING, 'john/' . $this->fileInfoNotSharedPdf->getPath(), 
                 '/test/path/to/file_OCR.pdf', '/tmp/ocr_randomTempFileName', OcrConstants::OCRmyPDF, $this->userId, 
-                false, $this->fileInfoSharedPdf->getName(), null);
+                false, $this->fileInfoSharedPdf->getName(), null, false);
         $job2 = new OcrJob(OcrConstants::STATUS_PENDING, 'notJohn/' . $this->fileInfoSharedPdf->getPath(), 
                 '/test/path/to/file_OCR.pdf', '/tmp/ocr_randomTempFileName', OcrConstants::OCRmyPDF, $this->userId, 
-                false, $this->fileInfoSharedPdf->getName(), null);
+                false, $this->fileInfoSharedPdf->getName(), null, false);
         $job3 = new OcrJob(OcrConstants::STATUS_PENDING, 'notJohn/' . $this->fileInfoSharedPng->getPath(), 
                 '/test/path/to/file_OCR.txt', '/tmp/ocr_randomTempFileName.txt', OcrConstants::TESSERACT, $this->userId, 
-                false, $this->fileInfoSharedPng->getName(), null);
+                false, $this->fileInfoSharedPng->getName(), null, false);
         $this->redisServiceMock->expects($this->exactly(3))
             ->method('sendJob')
             ->withConsecutive([
@@ -267,7 +332,7 @@ class JobServiceTest extends TestCase {
                 $job3,
                 []
         ]);
-        $result = $this->cut->process($languages, $files);
+        $result = $this->cut->process($languages, $files, $replace);
         $this->assertEquals('PROCESSING', $result);
     }
 
@@ -287,6 +352,7 @@ class JobServiceTest extends TestCase {
                 'deu',
                 'deu-frak'
         ];
+        $replace = false;
         $this->appConfigServiceMock->expects($this->once())
             ->method('getAppValue')
             ->with('languages')
@@ -314,16 +380,22 @@ class JobServiceTest extends TestCase {
             ->will($this->onConsecutiveCalls(false, true, true));
         $this->fileServiceMock->expects($this->exactly(3))
             ->method('buildTarget')
-            ->withConsecutive([
-                $this->fileInfoNotSharedPdf,
-                false
-        ], [
-                $this->fileInfoSharedPdf,
-                true
-        ], [
-                $this->fileInfoSharedPng,
-                true
-        ])
+            ->withConsecutive(
+                [
+                        $this->fileInfoNotSharedPdf,
+                        false,
+                        false
+                ], 
+                [
+                        $this->fileInfoSharedPdf,
+                        true,
+                        false
+                ], 
+                [
+                        $this->fileInfoSharedPng,
+                        true,
+                        false
+                ])
             ->will(
                 $this->onConsecutiveCalls('/test/path/to/file_OCR.pdf', '/test/path/to/file_OCR.pdf', 
                         '/test/path/to/file_OCR.txt'));
@@ -375,13 +447,13 @@ class JobServiceTest extends TestCase {
         // step out from tempfile creation
         $job1 = new OcrJob(OcrConstants::STATUS_PENDING, 'john/' . $this->fileInfoNotSharedPdf->getPath(), 
                 '/test/path/to/file_OCR.pdf', '/tmp/ocr_randomTempFileName', OcrConstants::OCRmyPDF, $this->userId, 
-                false, $this->fileInfoNotSharedPdf->getName(), null);
+                false, $this->fileInfoNotSharedPdf->getName(), null, false);
         $job2 = new OcrJob(OcrConstants::STATUS_PENDING, 'notJohn/' . $this->fileInfoSharedPdf->getPath(), 
                 '/test/path/to/file_OCR.pdf', '/tmp/ocr_randomTempFileName', OcrConstants::OCRmyPDF, $this->userId, 
-                false, $this->fileInfoSharedPdf->getName(), null);
+                false, $this->fileInfoSharedPdf->getName(), null, false);
         $job3 = new OcrJob(OcrConstants::STATUS_PENDING, 'notJohn/' . $this->fileInfoSharedPng->getPath(), 
                 '/test/path/to/file_OCR.txt', '/tmp/ocr_randomTempFileName.txt', OcrConstants::TESSERACT, $this->userId, 
-                false, $this->fileInfoSharedPng->getName(), null);
+                false, $this->fileInfoSharedPng->getName(), null, false);
         $this->redisServiceMock->expects($this->exactly(3))
             ->method('sendJob')
             ->withConsecutive([
@@ -394,7 +466,7 @@ class JobServiceTest extends TestCase {
                 $job3,
                 $languages
         ]);
-        $result = $this->cut->process($languages, $files);
+        $result = $this->cut->process($languages, $files, $replace);
         $this->assertEquals('PROCESSING', $result);
     }
 
@@ -417,6 +489,7 @@ class JobServiceTest extends TestCase {
         $languages = [
                 'any'
         ];
+        $replace = false;
         $this->appConfigServiceMock->expects($this->once())
             ->method('getAppValue')
             ->with('languages')
@@ -456,7 +529,7 @@ class JobServiceTest extends TestCase {
             ->method('tempnamWrapper')
             ->with('/tmp', OcrConstants::TEMPFILE_PREFIX)
             ->will($this->throwException(new NotFoundException('Temp file cannot be created.')));
-        $this->cut->process($languages, $files);
+        $this->cut->process($languages, $files, $replace);
     }
 
     /**
@@ -478,6 +551,7 @@ class JobServiceTest extends TestCase {
         $languages = [
                 'any'
         ];
+        $replace = false;
         $this->appConfigServiceMock->expects($this->once())
             ->method('getAppValue')
             ->with('languages')
@@ -522,7 +596,7 @@ class JobServiceTest extends TestCase {
             ->with('/tmp/ocr_randomTempFileName')
             ->willThrowException(
                 new NotFoundException('Cannot delete temporary file during temp file creation for Tesseract.'));
-        $this->cut->process($languages, $files);
+        $this->cut->process($languages, $files, $replace);
     }
 
     /**
@@ -544,6 +618,7 @@ class JobServiceTest extends TestCase {
         $languages = [
                 'any'
         ];
+        $replace = false;
         $this->appConfigServiceMock->expects($this->once())
             ->method('getAppValue')
             ->with('languages')
@@ -590,7 +665,7 @@ class JobServiceTest extends TestCase {
             ->method('touchWrapper')
             ->with('/tmp/ocr_randomTempFileName.txt')
             ->willThrowException(new NotFoundException('Cannot create temporary file for Tesseract.'));
-        $this->cut->process($languages, $files);
+        $this->cut->process($languages, $files, $replace);
     }
 
     /**
@@ -612,6 +687,7 @@ class JobServiceTest extends TestCase {
         $languages = [
                 'any'
         ];
+        $replace = false;
         $this->appConfigServiceMock->expects($this->once())
             ->method('getAppValue')
             ->with('languages')
@@ -661,7 +737,7 @@ class JobServiceTest extends TestCase {
             ->method('chmodWrapper')
             ->with('/tmp/ocr_randomTempFileName.txt', 0600)
             ->willThrowException(new NotFoundException('Cannot set permissions to temporary file for Tesseract.'));
-        $this->cut->process($languages, $files);
+        $this->cut->process($languages, $files, $replace);
     }
 
     /**
@@ -703,7 +779,7 @@ class JobServiceTest extends TestCase {
     public function testDeleteJob() {
         $job = new OcrJob(OcrConstants::STATUS_FAILED, 'john/' . $this->fileInfoNotSharedPdf->getPath(), 
                 '/test/path/to/file_OCR.pdf', '/tmp/ocr_randomTempFileName', OcrConstants::OCRmyPDF, $this->userId, 
-                false, $this->fileInfoNotSharedPdf->getName(), 'testLog');
+                false, $this->fileInfoNotSharedPdf->getName(), 'testLog', true);
         $job->setId(1);
         $this->jobMapperMock->expects($this->once())
             ->method('find')
@@ -722,6 +798,7 @@ class JobServiceTest extends TestCase {
         $this->assertNull($result->getTempFile());
         $this->assertNull($result->getType());
         $this->assertNull($result->getUserId());
+        $this->assertNull($result->getReplace());
         $this->assertEquals(1, $result->getId());
         $this->assertEquals($this->fileInfoNotSharedPdf->getName(), $result->getOriginalFilename());
     }
@@ -729,11 +806,11 @@ class JobServiceTest extends TestCase {
     public function testGetAllJobsForUser() {
         $job1 = new OcrJob(OcrConstants::STATUS_FAILED, 'john/' . $this->fileInfoNotSharedPdf->getPath(), 
                 '/test/path/to/file_OCR.pdf', '/tmp/ocr_randomTempFileName', OcrConstants::OCRmyPDF, $this->userId, 
-                false, $this->fileInfoNotSharedPdf->getName(), 'testLog');
+                false, $this->fileInfoNotSharedPdf->getName(), 'testLog', true);
         $job1->setId(1);
         $job2 = new OcrJob(OcrConstants::STATUS_PENDING, 'notJohn/' . $this->fileInfoSharedPng->getPath(), 
                 '/test/path/to/file_OCR.txt', '/tmp/ocr_randomTempFileName.txt', OcrConstants::TESSERACT, $this->userId, 
-                true, $this->fileInfoSharedPng->getName(), null);
+                true, $this->fileInfoSharedPng->getName(), null, false);
         $job2->setId(2);
         $this->jobMapperMock->expects($this->once())
             ->method('findAll')
@@ -747,6 +824,7 @@ class JobServiceTest extends TestCase {
         $this->assertEquals($this->fileInfoNotSharedPdf->getName(), $result[0]->getOriginalFilename());
         $this->assertEquals(OcrConstants::STATUS_FAILED, $result[0]->getStatus());
         $this->assertEquals('testLog', $result[0]->getErrorLog());
+        $this->assertTrue($result[0]->getReplace());
         $this->assertNull($result[0]->getErrorDisplayed());
         $this->assertNull($result[0]->getSource());
         $this->assertNull($result[0]->getTarget());
@@ -756,6 +834,7 @@ class JobServiceTest extends TestCase {
         $this->assertEquals(2, $result[1]->getId());
         $this->assertEquals($this->fileInfoSharedPng->getName(), $result[1]->getOriginalFilename());
         $this->assertEquals(OcrConstants::STATUS_PENDING, $result[1]->getStatus());
+        $this->assertFalse($result[1]->getReplace());
         $this->assertNull($result[1]->getErrorLog());
         $this->assertNull($result[1]->getErrorDisplayed());
         $this->assertNull($result[1]->getSource());
@@ -784,11 +863,11 @@ class JobServiceTest extends TestCase {
     public function testCheckForFinishedJobsSuccessfully() {
         $job1 = new OcrJob(OcrConstants::STATUS_PENDING, 'john/' . $this->fileInfoNotSharedPdf->getPath(), 
                 '/test/path/to/file_OCR.pdf', '/tmp/ocr_randomTempFileName', OcrConstants::OCRmyPDF, $this->userId, 
-                false, $this->fileInfoNotSharedPdf->getName(), null);
+                false, $this->fileInfoNotSharedPdf->getName(), null, true);
         $job1->setId(1);
         $job2 = new OcrJob(OcrConstants::STATUS_PENDING, 'notJohn/' . $this->fileInfoSharedPng->getPath(), 
                 '/test/path/to/file_OCR.txt', '/tmp/ocr_randomTempFileName.txt', OcrConstants::TESSERACT, $this->userId, 
-                false, $this->fileInfoSharedPng->getName(), null);
+                false, $this->fileInfoSharedPng->getName(), null, true);
         $job2->setId(2);
         $redisFinishedJobs = [
                 '{"id":1,"error":false,"log":""}',
@@ -826,7 +905,7 @@ class JobServiceTest extends TestCase {
     public function testHandleProcessedFileNotExists() {
         $job1 = new OcrJob(OcrConstants::STATUS_PROCESSED, 'john/' . $this->fileInfoNotSharedPdf->getPath(), 
                 '/test/path/to/file_OCR.pdf', '/tmp/ocr_randomTempFileName', OcrConstants::OCRmyPDF, $this->userId, 
-                false, $this->fileInfoNotSharedPdf->getName(), null);
+                false, $this->fileInfoNotSharedPdf->getName(), null, true);
         $job1->setId(1);
         $this->jobMapperMock->expects($this->once())
             ->method('findAllProcessed')
@@ -851,13 +930,15 @@ class JobServiceTest extends TestCase {
     }
 
     public function testHandleProcessed() {
-        $job1 = new OcrJob(OcrConstants::STATUS_PROCESSED, 'john/' . $this->fileInfoNotSharedPdf->getPath(), 
-                '/test/path/to/file_OCR.pdf', '/tmp/ocr_randomTempFileName', OcrConstants::OCRmyPDF, $this->userId, 
-                false, $this->fileInfoNotSharedPdf->getName(), null);
+        $job1 = new OcrJob(OcrConstants::STATUS_PROCESSED, 
+                'john/files/test.pdf', '/test/path/to/file_OCR.pdf', 
+                '/tmp/ocr_randomTempFileName', OcrConstants::OCRmyPDF, $this->userId, false, 
+                $this->fileInfoNotSharedPdf->getName(), null, true);
         $job1->setId(1);
-        $job2 = new OcrJob(OcrConstants::STATUS_PROCESSED, 'notJohn/' . $this->fileInfoSharedPng->getPath(), 
-                '/test/path/to/file_OCR.txt', '/tmp/ocr_randomTempFileName.txt', OcrConstants::TESSERACT, $this->userId, 
-                false, $this->fileInfoSharedPng->getName(), null);
+        $job2 = new OcrJob(OcrConstants::STATUS_PROCESSED, 
+                'notJohn/files/test.png' . $this->fileInfoSharedPng->getPath(), '/test/path/to/file_OCR.txt', 
+                '/tmp/ocr_randomTempFileName.txt', OcrConstants::TESSERACT, $this->userId, false, 
+                $this->fileInfoSharedPng->getName(), null, false);
         $job2->setId(2);
         $this->jobMapperMock->expects($this->once())
             ->method('findAllProcessed')
@@ -882,6 +963,9 @@ class JobServiceTest extends TestCase {
                 $job2->getTempFile()
         ])
             ->will($this->onConsecutiveCalls('stream1', 'stream2'));
+        $this->viewMock->expects($this->exactly(1))
+            ->method('unlink')
+            ->with('/test.pdf');
         $this->viewMock->expects($this->exactly(2))
             ->method('file_put_contents')
             ->withConsecutive([
@@ -890,7 +974,7 @@ class JobServiceTest extends TestCase {
         ], [
                 $job2->getTarget(),
                 'stream2'
-        ]);
+        ])->will($this->onConsecutiveCalls(true, true));
         $this->jobMapperMock->expects($this->exactly(2))
             ->method('delete')
             ->withConsecutive([
@@ -906,17 +990,20 @@ class JobServiceTest extends TestCase {
                 $job2->getTempFile()
         ]);
         $result = $this->cut->handleProcessed();
-        $this->assertEquals([$job1, $job2], $result);
+        $this->assertEquals([
+                $job1,
+                $job2
+        ], $result);
     }
 
     public function testHandleFailed() {
         $job1 = new OcrJob(OcrConstants::STATUS_FAILED, 'john/' . $this->fileInfoNotSharedPdf->getPath(), 
                 '/test/path/to/file_OCR.pdf', '/tmp/ocr_randomTempFileName', OcrConstants::OCRmyPDF, $this->userId, 
-                false, $this->fileInfoNotSharedPdf->getName(), null);
+                false, $this->fileInfoNotSharedPdf->getName(), null, true);
         $job1->setId(1);
         $job2 = new OcrJob(OcrConstants::STATUS_FAILED, 'notJohn/' . $this->fileInfoSharedPng->getPath(), 
                 '/test/path/to/file_OCR.txt', '/tmp/ocr_randomTempFileName.txt', OcrConstants::TESSERACT, $this->userId, 
-                false, $this->fileInfoSharedPng->getName(), null);
+                false, $this->fileInfoSharedPng->getName(), null, true);
         $job2->setId(2);
         $this->jobMapperMock->expects($this->once())
             ->method('findAllFailed')
@@ -934,7 +1021,7 @@ class JobServiceTest extends TestCase {
                 $job2->getTempFile()
         ]);
         $job1Origin = $job1;
-        $job2Origin= $job2;
+        $job2Origin = $job2;
         $job1->setErrorDisplayed(true);
         $job2->setErrorDisplayed(true);
         $this->jobMapperMock->expects($this->exactly(2))
@@ -945,6 +1032,9 @@ class JobServiceTest extends TestCase {
                 $job2
         ]);
         $result = $this->cut->handleFailed();
-        $this->assertEquals([$job1Origin, $job2Origin], $result);
+        $this->assertEquals([
+                $job1Origin,
+                $job2Origin
+        ], $result);
     }
 }